+++ data/rentals.json
@@ -1,0 +1,7 @@
+[
+  {
+    "date" : "2020-05-05",
+    "book_index" : 0,
+    "person_index" : 0
+  }
+]
